refactoring and release createChangedInternal function
https://online.sbis.ru/opendoc.html?guid=bd852bf3-b3d0-4f26-b996-25d37b5b4005
